// apps/odm-downloader/src/app/constants/topic.enum.ts
export enum Topic {
  ParseODM = 'odm.parse',
  Cleanup = 'odm.cleanup',
  Assemble = 'odm.assemble',
  ProcessODM = 'odm.process',
  DownloadFile = 'odm.download.file',
  ParseLicense = 'odm.license.parse',
  DownloadImage = 'odm.download.image',
  DownloadLicense = 'odm.license.download',
  ReturnAudiobook = 'odm.return.audiobook',
}